refactor(frontend): extract useScheduleForm composable (W5)
Largest WorkflowDetail slice, smoke-gated: move the entire schedule tab
(form state + reactive scheduleForm + option lists + computeds
isScheduleOnline/timezoneOptions/environmentFilteredOptions + scheduleRules +
loadScheduleOptions/handleWorkerGroupChange/previewScheduleTimes/
syncScheduleForm/saveScheduleConfig/handleToggleSchedule + the activeTab
watch) into composables/useScheduleForm.js. workflow/activeTab injected;
buildDolphinConfigParams/loadWorkflowDetail/getErrorMessage lazy-wrapped
(stay component-owned, shared). scheduleOptionsLoaded returned so the
engine-switch/detail-load reset paths keep working.

Verified: WorkflowDetail mount smoke + full suite (27 files / 116 passed);
npm run build, npm run lint (0 errors). Verbatim move; WorkflowDetail.vue
2523 -> 2240.
